osn_final/project-main/tests/requirements.txt
pytest>=7.0.0
pytest-cov>=3.0.0
pytest-xdist>=2.5.0
pytest-timeout>=2.1.0
requests>=2.28.0
python-dotenv>=0.21.0
